@@ -15,8 +15,8 @@
     "clean": "shx rm -rf dist lib-esm",
     "build": "tsc && rollup -c",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "storybook": "start-storybook -p 6006",
-    "build-storybook": "build-storybook",
+    "storybook": "storybook dev -p 6006",
+    "build-storybook": "storybook build",
     "prepare": "npm run clean && npm run build"
   },
   "repository": {
@@ -38,30 +38,40 @@
   },
   "devDependencies": {
     "@babel/core": "^7.10.2",
+    "@babel/preset-react": "^7.28.5",
+    "@babel/preset-typescript": "^7.28.5",
+    "@emotion/babel-plugin": "^11.13.5",
     "@rollup/plugin-node-resolve": "^8.0.1",
-    "@storybook/addon-actions": "^5.3.19",
-    "@storybook/addon-docs": "^5.3.19",
-    "@storybook/addon-links": "^5.3.19",
-    "@storybook/addon-viewport": "^5.3.19",
-    "@storybook/addons": "^5.3.19",
-    "@storybook/cli": "^5.3.19",
-    "@storybook/preset-typescript": "^3.0.0",
-    "@storybook/react": "^5.3.19",
-    "@types/react-dom": "^16.9.8",
+    "@storybook/addon-actions": "^8.4.6",
+    "@storybook/addon-docs": "^8.4.6",
+    "@storybook/addon-links": "^8.4.6",
+    "@storybook/addon-mdx-gfm": "^8.4.6",
+    "@storybook/addon-viewport": "^8.4.6",
+    "@storybook/addon-webpack5-compiler-babel": "^3.0.6",
+    "@storybook/addon-webpack5-compiler-swc": "^1.0.5",
+    "@storybook/cli": "^8.4.6",
+    "@storybook/react": "^8.4.6",
+    "@storybook/react-webpack5": "^8.4.6",
+    "@swc/plugin-emotion": "^12.0.0",
+    "@types/react": "^18.3.12",
+    "@types/react-dom": "^18.3.1",
     "babel-loader": "^8.1.0",
-    "rollup": "^2.14.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "remark-gfm": "^4.0.1",
+    "rollup": "^4.52.5",
     "rollup-plugin-commonjs": "^10.1.0",
     "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-peer-deps-external": "^2.2.2",
-    "rollup-plugin-typescript2": "^0.27.1",
+    "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-typescript2": "^0.36.0",
     "shx": "^0.3.2",
-    "storybook": "^5.3.19",
+    "storybook": "^8.4.6",
     "tslib": "^2.0.0",
-    "typescript": "^3.9.5"
+    "typescript": "^5.7.2"
   },
   "dependencies": {
-    "@emotion/core": "^10.0.28",
-    "@emotion/styled": "^10.0.27",
+    "@emotion/react": "^11.13.5",
+    "@emotion/styled": "^11.13.5",
     "prop-types": "^15.7.2"
   },
   "bit": {
@@ -70,5 +80,6 @@
     },
     "componentsDefaultDirectory": "components/{name}",
     "packageManager": "npm"
-  }
+  },
+  "packageManager": "yarn@1.22.22+sha512.a6b2f7906b721bba3d67d4aff083df04dad64c399707841b7acf00f6b133b7ac24255f2652fa22ae3534329dc6180534e98d17432037ff6fd140556e2bb3137e"
 }
