@@ -1,1 +1,2 @@
-export { default as Notification, NOTIFICATION_POSITION, NOTIFICATION_TYPE } from './Notification';
+export { default as Notification } from './Notification';
+export { NOTIFICATION_POSITION, NOTIFICATION_TYPE } from './types';
